design : 검색어 있을 시 홈화면 트렌딩 제거
